+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 *If you're reading this because you try make sense of some new API or a breaking change, you might also be interested in coming to the chat for explanations or guidance.*
 
+### next
+- better support of wide characters in InputField
+
 <a name="v0.19.1"></a>
 ### v0.19.1 - 2021-11-26
 - update Crossterm to 0.22.1 to fix some bugs